requirements.txt
fastapi
uvicorn
requests
aiohttp
beautifulsoup4
google-generativeai